TwainCamera: Prevent VC++ x64 warning.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@10952 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/TwainCamera/TwainDevice.cpp
@@ -1187,7 +1187,7 @@ public:
 					  {
 						 std::string sVal;
 						 getcurrent(&_cap, sVal);
-						 int len1 = min( valueSize, strlen(sVal.c_str()));
+						 size_t len1 = min( valueSize, strlen(sVal.c_str()));
 						 strncpy((char*)pvalue, sVal.c_str(), len1);
 						 break;
 					  }
